\name{don_dat}
\alias{don_dat}
\docType{data}
\title{
A sample dataset of human donor splice sites.
}
\description{
This dataset comprises 1000 donor splice site sequences, where each sequence is of length 20 with 10 at the exon end and 10 at the intron start excluding the conserved di-nucleotide GT at the begining of intron.
This dataset was randomly taken from true donor splice sites of HS3D dataset. 
}
\usage{data(don_dat)}

\references{
Pollastro P, Rampone S: HS3D: Homosapiens Splice Site Data Set. \emph{Nucleic Acids Res}. 2003, Molecular Biology Database Collection entry number 36; Annual Database Issue.
}
\examples{
data(don_dat)
}

